@@ -1,9 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Theme Helper Bundle for Contao Open Source CMS
  *
- * Copyright (C) 2019 pdir GmbH // pdir / digital agentur <https://pdir.de>
+ * Copyright (C) 2022 pdir GmbH // pdir / digital agentur <https://pdir.de>
  *
  * @package    pdir/contao-theme-helper-bundle
  * @link       https://github.com/pdir/contao-theme-helper-bundle
@@ -13,11 +15,17 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 namespace Pdir\ThemeHelperBundle\Backend;
 
-class Licence extends \BackendModule
+use Contao\BackendModule;
+use Contao\Environment;
+use Contao\Input;
+use Contao\ThemeModel;
+
+class Licence extends BackendModule
 {
-    protected $strTemplate = 'be_th_check_domain';
+    protected string $strTemplate = 'be_th_check_domain';
 
     /**
      * Generate the module
@@ -28,37 +36,37 @@ class Licence extends \BackendModule
         $this->Template->domainLabel = $GLOBALS['TL_LANG']['MSC']['th_insert_domain'];
         $this->Template->domainTip = $GLOBALS['TL_LANG']['MSC']['th_domain_tip'];
         $this->Template->buttonCheck = $GLOBALS['TL_LANG']['MSC']['th_button_check'];
-        $this->Template->shortCode = \Input::get('shortCode') ? : \Input::post('shortCode');
-        $this->Template->theme = \Input::get('theme');
+        $this->Template->shortCode = Input::get('shortCode') ? : Input::post('shortCode');
+        $this->Template->theme = Input::get('theme');
         $this->Template->message = null;
 
-        switch (\Input::get('act')) {
+        switch (Input::get('act')) {
             case 'checkDomain':
                 // make request to pdir api
                 $url = 'https://pdir.de/api/themes?';
 
                 $params = [
-                    'domain' => \Input::post('domain'),
-                    'ip' => \Environment::get('server'),
+                    'domain' => Input::post('domain'),
+                    'ip' => Environment::get('server'),
                 ];
 
-                $url .= http_build_query($params);
+                $url .= \http_build_query($params);
 
-                $ch = curl_init($url);
-                curl_setopt($ch, CURLOPT_HTTPHEADER, ['Accept:application/json, Content-Type:application/json']);
-                curl_setopt($ch, CURLOPT_CUSTOMREQUEST, 'GET');
-                curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
+                $ch = \curl_init($url);
+                \curl_setopt($ch, CURLOPT_HTTPHEADER, ['Accept:application/json, Content-Type:application/json']);
+                \curl_setopt($ch, CURLOPT_CUSTOMREQUEST, 'GET');
+                \curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
 
                 // response
-                $response = json_decode(curl_exec($ch), true);
+                $response = \json_decode(\curl_exec($ch), true);
 
-                if(isset($response['message']) && \Input::post('theme'))
+                if(isset($response['message']) && Input::post('theme'))
                 {
                     // Domain is registered
                     if($response['message'] == 'Domain registered' && $response['domain'])
                     {
                         /** @var ThemeModel $objTheme */
-                        $objTheme = \ThemeModel::findById(\Input::post('theme'));
+                        $objTheme = ThemeModel::findById(Input::post('theme'));
                         $objTheme->pdir_th_license_domain = $response['domain'];
                         $objTheme->save();
                     }
@@ -70,7 +78,7 @@ class Licence extends \BackendModule
 
                 break;
             default:
-                if(!\Input::get('shortCode'))
+                if(!Input::get('shortCode'))
                 {
                     $this->Template->readonly = true;
                 }
