@@ -1,9 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Theme Helper Bundle for Contao Open Source CMS
  *
- * Copyright (C) 2017 pdir / digital agentur <develop@pdir.de>
+ * Copyright (C) 2022 pdir GmbH / pdir / digital agentur <develop@pdir.de>
  *
  * @package    pdir/contao-theme-helper-bundle
  * @link       https://github.com/pdir/contao-theme-helper-bundle
@@ -13,12 +15,12 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-declare(strict_types=1);
+
 namespace Pdir\ThemeHelperBundle\EventListener;
 
 use Contao\ArticleModel;
 use Contao\ContentElement;
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\Events;
 use Contao\StringUtil;
 
@@ -30,7 +32,7 @@ use Contao\StringUtil;
 class InsertTagsListener
 {
     /**
-     * @var ContaoFrameworkInterface
+     * @var ContaoFramework
      */
     private $framework;
 
@@ -44,9 +46,9 @@ class InsertTagsListener
     /**
      * Constructor.
      *
-     * @param ContaoFrameworkInterface $framework
+     * @param ContaoFramework $framework
      */
-    public function __construct(ContaoFrameworkInterface $framework)
+    public function __construct(ContaoFramework $framework)
     {
         $this->framework = $framework;
     }
@@ -60,9 +62,9 @@ class InsertTagsListener
      */
     public function onReplaceInsertTags($tag)
     {
-        $elements = explode('::', $tag);
-        $key = strtolower($elements[0]);
-        if (in_array($key, $this->supportedTags, true)) {
+        $elements = \explode('::', $tag);
+        $key = \strtolower($elements[0]);
+        if (\in_array($key, $this->supportedTags, true)) {
             return $this->replaceThemeInsertTag($elements[1], $elements[2]);
         }
         return false;
