@@ -1,9 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * pdir theme helper bundle for Contao Open Source CMS
  *
- * Copyright (C) 2017 pdir / digital agentur <develop@pdir.de>
+ * Copyright (C) 2022 pdir GmbH / pdir / digital agentur <develop@pdir.de>
  *
  * @package    pdir/contao-theme-helper-bundle
  * @link       https://github.com/pdir/contao-theme-helper-bundle
@@ -13,7 +15,6 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-#declare(strict_types = 1);
 
 namespace Pdir\ThemeHelperBundle;
 
@@ -25,4 +26,8 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 class ThemeHelperBundle extends Bundle
 {
+    public function getPath(): string
+    {
+        return \dirname(__DIR__);
+    }
 }
